#33 Migrated subtitleSlice to subtitleReducer.ts
For consistency, subtitleSlice.ts was replaced with subtitleReducer, and separate actions
