+++ src/contentScript/component/actions/index.ts
@@ -1,1 +1,2 @@
 export * from './userActions';
+export * from './subtitleActions';
